apps/backend/requirements.txt
flask>=3.0.0
flask-cors>=5.0.0
google-adk>=1.18.0
google-genai>=1.45.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=8.0.0
flake8>=7.0.0
black>=24.0.0
mypy>=1.8.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
